# docs/swagger.yaml
basePath: /
definitions:
  transport.LoginResponse:
    properties:
      authToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  transport.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  transport.RegisterResponse:
    properties:
      authToken:
        type: string
      refreshToken:
        type: string
      user:
        $ref: '#/definitions/user.User'
    type: object
  user.User:
    description: User's login details
    properties:
      email:
        type: string
      id:
        type: string
      password:
        type: string
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Auth server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Log in a user by username and password
      parameters:
      - description: Login user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transport.LoginResponse'
      summary: Log in a user
      tags:
      - users
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user by username and password
      parameters:
      - description: User info
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/transport.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/transport.RegisterResponse'
      summary: Register a new user
      tags:
      - users
swagger: "2.0"